REF: Move Python conversions to separate file
